Refactor stat filter UI into StatFilterBuilder component
- Extracted the stat filter logic and UI from AdvancedFilters and ItemPacksDialog into a reusable StatFilterBuilder component. This reduces code duplication and centralizes the stat filter management, improving maintainability and consistency across the app.
